weather command: ApplicationCommand-only
